+++ src/main/java/com/ecommerceproject/modules/category/dto/CategoryResponseDto.java
@@ -1,4 +1,4 @@
-package com.ecommerceproject.modules.category.dto; // Cambia el paquete según la estructura de tu proyecto
+package com.ecommerceproject.modules.category.dto;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 @Setter
 @Getter
 public class CategoryResponseDto {
-    private Integer id;
+    private Long id;
 
     private String name;
